API

Vulnerability curve

ClimaCache.BrooksCorey — Method

BrooksCorey{FT}(vg::VanGenuchten{FT}) where {FT<:AbstractFloat}

A constructor for BrooksCorey to create BrooksCorey type soil from VanGenuchten type, given

vg VanGenuchten type soil water retention curve

SoilHydraulics.soil_ψ_25 — Function

This function calculates soil metric potential from soil retention curve type and soil volumetric water potential. The supported methods are

soil_ψ_25(bc, θ)

defined at /home/runner/work/SoilHydraulics.jl/SoilHydraulics.jl/src/vulnerability.jl:68.

soil_ψ_25(vg, θ)

defined at /home/runner/work/SoilHydraulics.jl/SoilHydraulics.jl/src/vulnerability.jl:98.

SoilHydraulics.soil_ψ_25 — Method

soil_ψ_25(bc::BrooksCorey{FT}, θ::FT) where {FT<:AbstractFloat}

Return the soil metric potential, given

bc BrooksCorey type structure

θ Soil volumetric water content (absolute value)

SoilHydraulics.soil_ψ_25 — Method

soil_ψ_25(vg::VanGenuchten{FT}, θ::FT) where {FT<:AbstractFloat}

Return the soil metric potential, given

vg VanGenuchten type structure

θ Soil volumetric water content (absolute value)

SoilHydraulics.relative_hydraulic_conductance — Function

This function return the hydraulic conductance ralative to maximum. The supported methods are

relative_hydraulic_conductance(bc, θ)

defined at /home/runner/work/SoilHydraulics.jl/SoilHydraulics.jl/src/vulnerability.jl:149.

relative_hydraulic_conductance(bc, ψ, ψ_25)

defined at /home/runner/work/SoilHydraulics.jl/SoilHydraulics.jl/src/vulnerability.jl:175.

relative_hydraulic_conductance(vg, θ)

defined at /home/runner/work/SoilHydraulics.jl/SoilHydraulics.jl/src/vulnerability.jl:200.

relative_hydraulic_conductance(vg, ψ, ψ_25)

defined at /home/runner/work/SoilHydraulics.jl/SoilHydraulics.jl/src/vulnerability.jl:226.

SoilHydraulics.relative_hydraulic_conductance — Method

relative_hydraulic_conductance(bc::BrooksCorey{FT}, θ::FT) where {FT<:AbstractFloat}

Return the relative hydraulic conductance of the soil, given

bc BrooksCorey type structure

θ Soil volumetric water content (absolute value)

SoilHydraulics.relative_hydraulic_conductance — Method

relative_hydraulic_conductance(bc::BrooksCorey{FT}, ψ::Bool, ψ_25::FT) where {FT<:AbstractFloat}

Return the relative hydraulic conductance of the soil, given

bc BrooksCorey type structure

ψ Bool to indicate that next parameter is potential

ψ_25 Soil metric potential at a reference temperature of 25 °C

SoilHydraulics.relative_hydraulic_conductance — Method

relative_hydraulic_conductance(vg::VanGenuchten{FT}, θ::FT) where {FT<:AbstractFloat}

Return the relative hydraulic conductance of the soil, given

vg VanGenuchten type structure

θ Soil volumetric water content (absolute value)

SoilHydraulics.relative_hydraulic_conductance — Method

relative_hydraulic_conductance(vg::VanGenuchten{FT}, ψ::Bool, ψ_25::FT) where {FT<:AbstractFloat}

Return the relative hydraulic conductance of the soil, given

vg VanGenuchten type structure

ψ Bool to indicate that next parameter is potential

ψ_25 Soil metric potential at a reference temperature of 25 °C